Replace the class extraction script (#11)
