+++ SFML.xml
@@ -1,4 +1,4 @@
 <entry>
-  <version>2.2.0.1</version>
+  <version>2.2.0.2</version>
   <url>https://raw.githubusercontent.com/tsoding/sfml-docset/master/SFML.tgz</url>
 </entry>
